refactor : 빌드 오류 해결
