+++ app/(anon)/@detail/steps/[step-number]/[detail]/_components/contents/DataGrid.tsx
@@ -2,19 +2,20 @@ import styles from './DataGrid.styles';
 import CircularIconBadge from '@/(anon)/_components/common/circularIconBadges/CircularIconBadge';
 
 interface DataGridProps {
-  data: Record<string, string>;
+  data: Array<{ left: string; right?: string }>;
 }
 
 const DataGrid = ({ data }: DataGridProps) => {
   return (
     <div>
-      {Object.entries(data).map(([key, value], index) => (
+      {data.map((item, index) => (
         <div key={index} className={styles.detailItem}>
-          <div className={styles.detailKey}>
-            {key}:
-          </div>
+          <div className={styles.detailKey}>{item.left}:</div>
           <div className={styles.detailValue}>
-            <CircularIconBadge type={value as 'match' | 'mismatch' | 'unchecked'} size="xsm" />
+            <CircularIconBadge
+              type={item.right as 'match' | 'mismatch' | 'unchecked'}
+              size='xsm'
+            />
           </div>
         </div>
       ))}
