+++ app/(anon)/@detail/steps/[step-number]/[detail]/_components/contents/List.tsx
@@ -2,14 +2,22 @@ import React from 'react';
 import styles from './List.styles';
 
 interface ListProps {
-  data: Record<string, string>;
+  data: Array<{ left: string; right?: string }>;
 }
 
 const List = ({ data }: ListProps) => {
   return (
     <div className={styles.container}>
       <div className={styles.content}>
-        List 컴포넌트 - data: {JSON.stringify(data)}
+        {data.map((item, index) => (
+          <div
+            key={index}
+            className='flex justify-between items-center py-2 border-b border-gray-200'
+          >
+            <span className='font-medium'>{item.left}</span>
+            {item.right && <span className='text-gray-600'>{item.right}</span>}
+          </div>
+        ))}
       </div>
     </div>
   );
