+++ app/(anon)/real-estate-data/_components/pageContent/PageContent.styles.ts
@@ -1,0 +1,5 @@
+export const styles = {
+  container: 'min-h-screen bg-brand-light-gray p-6',
+  wrapper: 'max-w-6xl mx-auto',
+  mainCard: 'bg-brand-white rounded-lg shadow-lg p-6',
+} as const;
