+++ app/(anon)/real-estate-data/_components/searchSection/SearchSection.styles.ts
@@ -1,0 +1,10 @@
+export const styles = {
+  header: 'mb-6',
+  title: 'text-2xl font-bold text-brand-black mb-4',
+  searchContainer: 'flex gap-4 items-center',
+  searchInput:
+    'flex-1 px-4 py-2 border border-brand-light-gray rounded-lg focus:ring-2 focus:ring-brand-blue focus:border-transparent',
+  resetButton:
+    'px-4 py-2 bg-brand-dark-gray text-brand-white rounded-lg hover:bg-brand-black transition-colors',
+  searchResult: 'text-sm text-brand-dark-gray mt-2',
+} as const;
